Add description to PDSMS desktop menu entry on Linux
This adds a short description to the desktop menu entry installed by the PDSMS Linux script. It also adds the program to the "Graphics" menu category along with "Development", which I found more appropriate. Made because I was bugged by the program being the only one without a description in my menu lol